requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-ckeditor==6.4.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.28.0
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
